Refactoring DriveInfo to work with BSD systems

+++ src/System.IO.FileSystem.DriveInfo/src/System/IO/DriveInfo.BSD.cs
@@ -47,7 +47,7 @@ namespace System.IO
             get
             {
                 Interop.libc.statfs data = Interop.libc.GetStatFsForDriveName(Name);
-                return (long)(data.f_bsize * data.f_bavail);
+                return (long)data.f_bsize * (long)data.f_bavail;
             }
         }
